pyproject.toml
[project]
name = "sparketl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jaydebeapi>=1.2.3",
    "jpype1>=1.6.0",
    "pyspark[connect]==3.5.5",
    "pyyaml>=6.0.3",
    "sqlparse>=0.5.5",
]
